assign2tester.c
This was the second assignment for operating systems. In this assignment we were to design a shell in C and have it prompt the user with a command line and track the prompts entered in the command line into a recent history. It has been a while since I've revisited this project and I remember there being a few issues but overall the program worked well.